Linter: add 'avoid_relative_imports'
This lint is necessary on codebases where there are often files being
moved around. The lack of consistent
'package:my.project.folder/mypackage.dart' makes it very painful to
detect when files are imported since not even intellij tells you that.

This lint is not supposed to be used for everyone since there are a lot
of valid cases for relative imports but I would prefer to have it as
warning or lints, not errors.